requirements.txt
amqp==5.3.1
annotated-types==0.7.0
APScheduler==3.11.2
async-timeout==5.0.1
babel==2.17.0
billiard==4.2.4
blinker==1.9.0
cachelib==0.13.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
croniter==6.0.0
dnspython==2.7.0
email-validator==2.3.0
exceptiongroup==1.3.1
Flask==3.1.2
Flask-APScheduler==1.13.1
Flask-Liquid==1.2.0
Flask-Mailman==1.1.1
Flask-Session==0.8.0
Flask-WTF==1.2.2
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.6.1
MarkupSafe==3.0.3
msgspec==0.20.0
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.0.1
requests==2.32.5
rq==1.13.0
six==1.17.0
SQLAlchemy==2.0.45
sqlmodel==0.0.29
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==1.26.6
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.4
WTForms==3.2.1
zipp==3.23.0
